@pdroalves 2025-01-06
Performance improvement on H100s is also quite impressive:

| Bench | timing ms | latency difference w.r.t previous impl |
|--------------------------------------|--------|---------|
| cuda::packing_compression::pack_u2   | 2.4167 | -37.42% |
| cuda::packing_compression::pack_u8   | 2.4250 | -32.89% |
| cuda::packing_compression::pack_u16  | 2.4298 | -32.18% |
| cuda::packing_compression::pack_u32  | 2.4338 | -32.20% |
| cuda::packing_compression::pack_u64  | 2.4452 | -32.26% |
| cuda::packing_compression::pack_u128 | 2.4713 | -32.88% |
| cuda::packing_compression::pack_u256 | 2.6083 | -61.40% |
| cuda::packing_compression::pack_u512 | 2.6287 | -75.61% |